pyproject.toml
[tool.poetry]
name = "hartford"
version = "0.1.0"
description = "Demo for Hartford Insurance"
authors = ["Rohan Singh <Rohan.Singh@ibm.com>"]

[tool.poetry.dependencies]
python = "^3.7.6"
fastapi = "^0.63.0"
kafka-python = "^2.0.2"
python-dotenv = "^0.17.0"
faker = "^7.0.1"
uvicorn = "^0.13.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
